+++ battleship_game/src/board_mechanics/piece.py
